line 142-143
est your container skills with PODESCAPE, an endless runner of a pod trying to escape from a project being deleted. Compete with fellow devs for the highest score. And don't be fooled by the retro feel...with the git link you can tinker under the hood with OpenShift on IBM Cloud, MongoDB, and OpenShift Container Storage.