code of conduct line 281
Code of Conduct:
TL;DR – Be kind and respectful.

We follow the Code of Conduct of other  events such as KubeCon. Similarly we are dedicated to providing a harassment-free experience for participants at all of our events, whether they are held in person or virtually. All event participants, whether they are attending an in-person event or a virtual event, are expected to behave in accordance with professional standards, with both this Code of Conduct as well as their respective employer’s policies governing appropriate workplace behavior and applicable laws.

<replaced>

The OpenShift Commons Gatherings bring together experts from all over the world
                  to discuss container technologies, best practices for cloud native application developers and the open source software projects that underpin the OpenShift
                  ecosystem. This Gathering will gather developers, DevOps professionals, and sysadmins together to explore the next steps in making container technologies successful and secure.