line 23 event footer (#1995)
The Hands-on Workshop will cover deploying microservices to an OpenShift cluster, viewing logs and metrics, and the use of an operator to provision IBM Cloud services. You'll get step-by-step guidance on Initializing an OpenShift application with a repository URL resulting in a Builder, Deployment, and Service. Services covered:  OpenShift, IBM Cloudant,  LogDNA and  IBM Cloud Monitoring with Sysdig.